Lib

Description

|

Documentation

parseFile :: String -> ([[Double]], Vector Double, Int)

parse the file and returns the training data, target values and number of vars

evalFitness

Arguments

:: Int
number of terms

-> [[Double]]
matrix X

-> Vector Double
vector y

-> [Bool]
variables of the poly

-> [Int]
exponents of the poly

-> Solution

Creates a fitness function

select :: Int -> Population -> StdGen -> (Population, StdGen)

Select the next generation

createRndSolution :: Int -> Fitness -> StdGen -> (Solution, StdGen)

Creates a random solution

crossover :: Fitness -> Population -> StdGen -> (Solution, StdGen)

Creates a new solution from the combination of two solutions

mutate :: Double -> Fitness -> Solution -> StdGen -> (Solution, StdGen)

Mutates a single solution

ga :: Int -> Int -> CreateSolution -> Crossover -> Mutation -> Fitness -> Selection -> StdGen -> (Population, StdGen)

Main function for Genetic Algorithm